Add initial unit tests for Results.vue
Why these changes are being introduced:
This backfills testing for the Results component. We had delayed
testing this component until we had a better handle on vue-test-utils,
and specifically how to mock API calls and routes.

Relevant ticket(s):
https://mitlibraries.atlassian.net/browse/DISCO-153

How this addresses that need:
This adds a few rudimentary unit tests for the Record component, using
mocked axios responses and mocked routes. The approach is similar to
our initial rests for the Record component.

Side effects to this change:
* The searchTimdex method is now async
* We no longer use function mode to pass the query param as a prop, as
it was unclear how to mock this in the tests
* The debug messages in Results have been removed
* We added RouterLinkStub to quiet the warnings about Item components
that render without links

Co-authored-by: Jeremy Prevost <jprevost@mit.edu>