Adds pagination to Results list
Why are these changes being introduced:

* Allowing users to look at more than the first set of results is
  important

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/DISCO-40

How does this address that need:

* This implements some basic logic to determine which pagination nav
  links are appropriate (next, previous, first, none)
* Each appropriate link will update the querystring and thus perform a
  new search

Document any side effects to this change:

* Some web pagination systems allow arbitrary jumping through results.
  At this time, it was determined that doesn't make a lot of sense as
  our results are only ordered by relevance and jumping past "more"
  relevant results to get to "less" relevant results is not a use case
  that makes sense to support. However, if we provide other sort orders,
  it may make more sense to implment more logic in the pagination to
  provide jumps (i.e. if sorted by data, jump a bunch of pages to skip
  the ranges you aren't interested in... or if sorted alphabetically
  jump to the "P" set)

* Removes default component border color

* Adds conditional rendering for Items

This uses the existing status loading flag to hide Item rendering until
the items are loaded.

It became particularly noticeble during pagination where the old records
were remaining and the user would be stuck to the bottom of the viewport
and not see the loading message and would have to scroll up when the
items loaded in.

* Adds scroll behavior to router

If we have a savedPosition or pass a url hash, use that position,
otherwise use the top of the page.

https://next.router.vuejs.org/guide/advanced/scroll-behavior.html

* Removes unused config variable

Eventually we will want this. Right now we do nothing with it though so
having it around just makes things misleading as changing it has no
positive effect but will break pagination