Implement faceting on Results list
Why these changes are being introduced:

Facets are a high priority for the new discovery UI.

Relevant ticket(s):

* Closes: https://mitlibraries.atlassian.net/browse/DISCO-41
* Relates to: https://mitlibraries.atlassian.net/browse/DISCO-210
* Relates to: https://mitlibraries.atlassian.net/browse/DISCO-215

How this addresses that need:

This adds a v-model to the Facet component, such that whenever a checkbox for
a given facet is check or unchecked, that facet is added to or removed from the
v-model. The component watches the v-model, and when the data changes, it calls
a method that updates the URL params accordingly.

Side effects of this change:

* There are some minor changes to the Results component around how we construct
the query string, because TIMDEX uses the brackets syntax for arrays.
* Since input data isn't persisted, checked boxes reset when the page is
refreshed, or when browser history changes. This work will be tracked in
DISCO-210.
* More comprehensive testing is required. This work will be tracked in
DISCO-215.