Add route reactivity to Facet.vue
Why these changes are being introduced:

Currently, Facet.vue does not react to changes in the route. This
means that the v-model resets whenever the page reloads, and does
not change when the browser history changes.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/DISCO-210

How this addresses that need:

This adds a route watcher that updates facets from the URL query
params, so whenever the route changes, the v-model updates accordingly.

Side effects of this change:

Some minor tweaks to facet.spec.js, in addition to coverage for
the new functionality.