Fix bug in Facet.vue that breaks pagination
Why these changes are being introduced:

We want to return users to the first page of results when facets are
applied or removed. However, ince we added the updateFacetsFromURL()
method, it is impossible to access any results beyond page 1. This is
because updateFacetsFromURL() mutates the v-model, which calls
the applyFacets() method, which in turn reverts to page 1.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/DISCO-240

How this addresses that need:

Adds a boolean data property to track whether updateFacetsFromURL()
has been run. If it hasn't, applyFacets() will reset the page to 1,
then toggle that property to true.

Side effects of this change:

* Miscellaneous cleanup, including:
    * Fixing misnamed test
    * Updating watcher syntax in Results.vue
    * Using consistent syntax when accessing wrapper data
* The new syntax in the Results route watcher causes a small drop in
code coverage. In fact, this was never covered, and I'd argue that it
shouldn't be, since we would effectively be testing the framework.
* We will need to write our first integration tests to confirm the
debugged behavior when clicking the pagination buttons. I've begun
working on this and will continue it in a separate ticket/branch.